+++ templates/database/seeders/FileSeeder.php
@@ -18,14 +18,26 @@ class FileSeeder extends Seeder
             'scope' => 'filegroup',
             'name' => 'public',
             'order' => '10',
-            'shortname' => 'P'
+            'shortname' => 'PUB'
         ]);
         DB::table('sproperties')->insert([
             'id' => 1102,
             'scope' => 'filegroup',
             'name' => 'private',
             'order' => '20',
-            'shortname' => 'P'
+            'shortname' => 'PRIV'
+        ]);
+        DB::table('sproperties')->insert([
+            'id' => 1103, 'scope' => 'filegroup', 'name' => 'Audio file', 'order' => '30', 'shortname' => 'AUDIO'
+        ]);
+        DB::table('sproperties')->insert([
+            'id' => 1104, 'scope' => 'filegroup', 'name' => 'Video file', 'order' => '40', 'shortname' => 'VIDEO'
+        ]);
+        DB::table('sproperties')->insert([
+            'id' => 1105, 'scope' => 'filegroup', 'name' => 'Image file', 'order' => '50', 'shortname' => 'IMG'
+        ]);
+        DB::table('sproperties')->insert([
+            'id' => 1106, 'scope' => 'filegroup', 'name' => 'Document file', 'order' => '60', 'shortname' => 'DOC'
         ]);
     }
 }